consolidate temp fxns to BLS_QCEW
