+++ flowsa/methods/flowbysectormethods/SEEA_2016_v1.yaml
@@ -64,7 +64,7 @@ sources_to_cache:
     selection_fields:
       Class: Employment
     year: 2012
-    clean_fba: !script_function:temp_data_source_functions clean_qcew
+    clean_fba: !script_function:BLS_QCEW clean_qcew
   BTS_TSA:
     year: 2016
     selection_fields:
