+++ flowsa/methods/flowbysectormethods/SEEA_2017_kg_v1.yaml
@@ -80,7 +80,7 @@ sources_to_cache:
     selection_fields:
       Class: Employment
     year: 2012
-    clean_fba: !script_function:temp_data_source_functions clean_qcew
+    clean_fba: !script_function:BLS_QCEW clean_qcew
   BTS_TSA:
     year: 2017
     selection_fields:
